double fix

+++ DataLayer/DTOs/Product/v2_ProductDTO.cs
@@ -7,7 +7,7 @@ namespace AuthReadyAPI.DataLayer.DTOs.Product
         public string? stripeId { get; set; }
         public string name { get; set; }
         public string? description { get; set; }
-        public long default_price { get; set; }
+        public double default_price { get; set; }
         public int quantity { get; set; }
         public Boolean livemode = false;
         public string? package_dimensions { get; set; }
